Factor argument and return values out of ffidl_cif
The ffidl_cif should function as a wrapper over the FFI library types,
augmenting it with Ffidl-specific information (e.g., in which contexts
is a type supported).  This information is useful for both callouts and
callbacks.

Details relating to the invocation of a funcion, as well as the storage
for is arguments is relevant only to the callouts.

As a consequence of this change, we can check libffi's raw API support
separately for callouts and callbacks.